shorter but clearer
I find it is no necessary to build a function to get the transform result because this work can be done within two lines code. Therefore, I combine them into a anonymous object in the output line. In the meanwhile, other symbols are put in the format output line, which is more suitable and shorter.